Documentacion de todas las clases en controllers. Creación del metodo constructor en cada clase de controllers. Refactor en las validaciones de id

+++ src/main/java/com/bitala/apiprueba/controllers/AnomaliasController.java
@@ -8,36 +8,54 @@ import com.bitala.apiprueba.repository.IAnomaliasRepository;
 
 import java.util.List;
 
+/**
+ * API MANTENIMIENTOS - BITALA
+ * @AUTHOR ERICK GAMEZ
+ * CONTROLLER - ANOMALIAS
+ * 
+ * version 1.0
+ */
+
+//Indica que la clase es un controlador
 @RestController
+//Ruta para acceder a Anomalias
 @RequestMapping("/api/anomalia")
 public class AnomaliasController {
     
-    //Anotación para la inyección de dependencias
+    //Anotación para inyección de dependencias
     @Autowired
     private IAnomaliasRepository anomaliasRepository;
 
-    //Objeto Anomalias
-    Anomalias anomalia;
+    //Constructor para inyeccion de dependencias
+    public AnomaliasController(IAnomaliasRepository anomaliasRepository) {
+        this.anomaliasRepository = anomaliasRepository;
+    }
 
+    //Retorna una lista de todos los elementos de Anomalias
     @GetMapping
     public List<Anomalias> allAnomalias(){
         return anomaliasRepository.findAll();
     }
 
+    //Busca una Anomalia por id
     @GetMapping("/{id}")
     public Anomalias findById(@PathVariable("id") Long id){
-        return anomaliasRepository.findById(id).orElse(null);
+        if (anomaliasRepository.existsById(id)) return anomaliasRepository.findById(id).orElse(null);
+        else return null;
     }
 
+    //Agrega una nueva Anomalia
     @PostMapping
     public Anomalias createAnomalia(@RequestBody Anomalias anomalia){
         return anomaliasRepository.save(anomalia);
     }
 
+    //Modifica una Anomalia por id
     @PutMapping("/{id}")
     public Anomalias updateAnomalia(@PathVariable Long id, @RequestBody Anomalias anomaliaData){
-        anomalia = anomaliasRepository.findById(id).orElse(null);
-        if (anomalia != null) {
+        Anomalias anomalia = new Anomalias();
+
+        if (anomaliasRepository.existsById(id)) {
             anomalia.setIdUnidad(anomaliaData.getIdUnidad());
             anomalia.setIdChofer(anomaliaData.getIdChofer());
             anomalia.setIdMantenimiento(anomaliaData.getIdMantenimiento());
@@ -46,13 +64,12 @@ public class AnomaliasController {
             anomalia.setFecha(anomaliaData.getFecha());
             anomalia.setEstatus(anomaliaData.getEstatus());
             return anomaliasRepository.save(anomalia);
-        } else {
-            return null;
-        }
+        } else return null;
     }
 
+    //Elimina una Anomalia por id
     @DeleteMapping("/{id}")
     public void deleteAnomalia(@PathVariable("id") Long id){
-        anomaliasRepository.deleteById(id);
+        if(anomaliasRepository.existsById(id)) anomaliasRepository.deleteById(id);
     }
 }
