+++ src/main/java/com/bitala/apiprueba/controllers/CheckCategoriaController.java
@@ -8,7 +8,17 @@ import com.bitala.apiprueba.repository.ICheckCategoriaRepository;
 
 import java.util.List;
 
+/**
+ * API MANTENIMIENTOS - BITALA
+ * @AUTHOR ERICK GAMEZ
+ * CONTROLLER - CHECKCATEGORIA
+ * 
+ * version 1.0
+ */
+
+//Indica que la clase es un controlador
 @RestController
+//Ruta para acceder a CheckCategoria
 @RequestMapping("/api/check-categoria")
 public class CheckCategoriaController {
 
@@ -16,37 +26,44 @@ public class CheckCategoriaController {
     @Autowired
     private ICheckCategoriaRepository checkCategoriaRepository;
 
-    //Objeto CheckCategoria
-    CheckCategoria checkCategoria;
+    //Constructor para inyección de dependencias
+    public CheckCategoriaController(ICheckCategoriaRepository checkCategoriaRepository) {
+        this.checkCategoriaRepository = checkCategoriaRepository;
+    }
 
+    //Retorna una lista de todos los elementos de CheckCategoria
     @GetMapping
     public List<CheckCategoria> allCategorias(){
         return checkCategoriaRepository.findAll();
     }
 
+    //Busca un CheckCategoria por id
     @GetMapping("/{id}")
     public CheckCategoria findById(@PathVariable("id") Long id){
-        return checkCategoriaRepository.findById(id).orElse(null);
+        if (checkCategoriaRepository.existsById(id)) return checkCategoriaRepository.findById(id).orElse(null);
+        else return null;
     }
 
+    //Agrega un nuevo CheckCategoria
     @PostMapping
     public CheckCategoria createCategoria(@RequestBody CheckCategoria categoria){
         return checkCategoriaRepository.save(categoria);
     }
 
+    //Modifica un CheckCategoria por id
     @PutMapping("/{id}")
     public CheckCategoria updateCategoria(@PathVariable Long id, @RequestBody CheckCategoria categoriaData) {
-        checkCategoria = checkCategoriaRepository.findById(id).orElse(null);
-        if(checkCategoria != null){
+        CheckCategoria checkCategoria = new CheckCategoria();
+
+        if(checkCategoriaRepository.existsById(id)){
             checkCategoria.setNombre(categoriaData.getNombre());
             return checkCategoriaRepository.save(checkCategoria);
-        } else {
-            return null;
-        }
+        } else return null;
     }
 
+    //Elimina un CheckCategoria por id
     @DeleteMapping("/{id}")
 	public void deleteCategoria(@PathVariable("id") Long id) {
-		checkCategoriaRepository.deleteById(id);
+        if (checkCategoriaRepository.existsById(id)) checkCategoriaRepository.deleteById(id);
 	}
 }
