+++ src/main/java/com/bitala/apiprueba/controllers/MantenimientoController.java
@@ -8,7 +8,17 @@ import com.bitala.apiprueba.repository.IMantenimientoRepository;
 
 import java.util.List;
 
+/**
+ * API MANTENIMIENTOS - BITALA
+ * @AUTHOR ERICK GAMEZ
+ * CONTROLLER - MANTENIMIENTO
+ * 
+ * version 1.0
+ */
+
+//Indica que la clase es un controlador
 @RestController
+//Ruta para acceder a Mantenimiento
 @RequestMapping("/api/mantenimiento")
 public class MantenimientoController {
     
@@ -16,41 +26,48 @@ public class MantenimientoController {
     @Autowired
     private IMantenimientoRepository mantenimientoRepository;
 
-    //Objeto Mantenimiento
-    Mantenimiento mantenimiento;
+    //Constructor para inyección de dependencias
+    public MantenimientoController(IMantenimientoRepository mantenimientoRepository) {
+        this.mantenimientoRepository = mantenimientoRepository;
+    }
 
+    //Retorna una lista de todos los elementos de Mantenimiento
     @GetMapping
     public List<Mantenimiento> allMantenimientos(){
         return mantenimientoRepository.findAll();
     }
 
+    //Busca un Mantenimiento por id
     @GetMapping("/{id}")
     public Mantenimiento findById(@PathVariable("id") Long id){
-        return mantenimientoRepository.findById(id).orElse(null);
+        if(mantenimientoRepository.existsById(id)) return mantenimientoRepository.findById(id).orElse(null);
+        else return null;
     }
 
+    //Agrega un nuevo Mantenimiento
     @PostMapping
     public Mantenimiento createMantenimiento(@RequestBody Mantenimiento mantenimiento){
         return mantenimientoRepository.save(mantenimiento);
     }
 
+    //Modifica un Mantenimiento por id
     @PutMapping("/{id}")
     public Mantenimiento updateMantenimiento(@PathVariable Long id, @RequestBody Mantenimiento manteData){
-        mantenimiento = mantenimientoRepository.findById(id).orElse(null);
-        if (mantenimiento != null) {
+        Mantenimiento mantenimiento = new Mantenimiento();
+
+        if (mantenimientoRepository.existsById(id)) {
             mantenimiento.setIdUnidad(manteData.getIdUnidad());
             mantenimiento.setFechaSolicitada(manteData.getFechaSolicitada());
             mantenimiento.setFechaEntrega(manteData.getFechaEntrega());
             mantenimiento.setEstatus(manteData.getEstatus());
             mantenimiento.setObservacion(manteData.getObservacion());
             return mantenimientoRepository.save(mantenimiento);
-        } else {
-            return null;
-        }
+        } else return null;
     }
 
+    //Elimina un Mantenimiento por id
     @DeleteMapping("/{id}")
     public void deleteMantenimiento(@PathVariable("id") Long id){
-        mantenimientoRepository.deleteById(id);
+        if(mantenimientoRepository.existsById(id)) mantenimientoRepository.deleteById(id);
     }
 }
